+++ packages/api/src/functions/upload.ts
@@ -24,11 +24,15 @@ const streamToFs = (metadata: typeof FileMetadata, context: Context) => {
     ? metadata.filename.split('.').pop()
     : 'bin'
 
-  if( !context.apiConfig.storage?.tempFs ) {
+  const tmpPath = context.apiConfig.storage
+    ? context.apiConfig.storage.tempFs || context.apiConfig.storage.fs
+    : null
+
+  if( !tmpPath ) {
     throw new Error()
   }
 
-  const absolutePath = path.join(context.apiConfig.storage.tempFs, `${filenameHash}.${extension}`)
+  const absolutePath = path.join(tmpPath, `${filenameHash}.${extension}`)
 
   return new Promise<string>((resolve, reject) => {
     const stream = createWriteStream(absolutePath)